LayoutTests/fast/scrolling/scrollIntoView-on-option-expected.txt


PASS scroller.scrollTop is within 10 of 790
PASS listbox.scrollTop is within 2 of 52
PASS successfullyParsed is true

TEST COMPLETE
